Separate router links into app nav (green) and site nav (black) headers
Furthers the process of mimicking the form and function of the PSU website. New green nav bar contains links to the login page and home page while the links to individual vue components stay in the black nav bar.

In the future the login link should be replaced with a logout button.