Create new page-component class
Each component that represented one of the original 3 pages had the same main-content id attribute that had their shared css, obviously a bad idea. So the css for the components was revamped and placed in a page-component class—these components are made of content that use to be in the old pages after all.

These same components also had a bunch of empty classes, so this commit removes those too.